Add files via upload
I spent way too long trying to get my data to load and trying to make figures with the one file I did get to load and I was afraid that I would not figure it out in time for class, so I am going to use the figures I made in R for the class presentation today. This is the code I used to generate those figures.